+++ audioswitch/src/main/java/com/twilio/audioswitch/selection/AudioDeviceSelector.kt
@@ -5,7 +5,7 @@ import android.content.Context
 import android.media.AudioManager
 import com.twilio.audioswitch.BuildConfig
 import com.twilio.audioswitch.android.BuildWrapper
-import com.twilio.audioswitch.android.LogWrapper
+import com.twilio.audioswitch.android.Logger
 import com.twilio.audioswitch.bluetooth.BluetoothHeadsetConnectionListener
 import com.twilio.audioswitch.bluetooth.BluetoothHeadsetManager
 import com.twilio.audioswitch.selection.AudioDevice.BluetoothHeadset
@@ -42,7 +42,7 @@ class AudioDeviceSelector {
      */
     constructor(context: Context) {
         val audioManager = context.getSystemService(Context.AUDIO_SERVICE) as AudioManager
-        val logger = LogWrapper()
+        val logger = Logger()
         val audioDeviceManager =
                 AudioDeviceManager(context,
                         logger,
@@ -57,7 +57,7 @@ class AudioDeviceSelector {
     }
 
     internal constructor(
-        logger: LogWrapper,
+        logger: Logger,
         audioDeviceManager: AudioDeviceManager,
         wiredHeadsetReceiver: WiredHeadsetReceiver,
         headsetManager: BluetoothHeadsetManager?
@@ -68,7 +68,7 @@ class AudioDeviceSelector {
         this.bluetoothHeadsetManager = headsetManager
     }
 
-    private var logger: LogWrapper = LogWrapper()
+    private var logger: Logger = Logger()
     private val audioDeviceManager: AudioDeviceManager
     private val wiredHeadsetReceiver: WiredHeadsetReceiver
     internal var audioDeviceChangeListener: AudioDeviceChangeListener? = null
@@ -221,6 +221,23 @@ class AudioDeviceSelector {
             enumerateDevices()
         }
     }
+
+    /**
+     * A property to configure AudioSwitch logging behavior. AudioSwitch logging is disabled by
+     * default.
+     */
+    var loggingEnabled: Boolean
+        /**
+         * Returns `true` if logging is enabled. Returns `false` by default.
+         */
+        get() = logger.loggingEnabled
+
+        /**
+         * Toggle whether logging is enabled.
+         */
+        set(value) {
+            logger.loggingEnabled = value
+        }
 
     /**
      * Retrieves the selected [AudioDevice] from [AudioDeviceSelector.selectDevice].