+++ audioswitch/src/test/java/com/twilio/audioswitch/wired/WiredHeadsetReceiverTest.kt
@@ -7,7 +7,7 @@ import com.nhaarman.mockitokotlin2.isA
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
-import com.twilio.audioswitch.android.LogWrapper
+import com.twilio.audioswitch.android.Logger
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.CoreMatchers.nullValue
@@ -18,7 +18,7 @@ import org.junit.Test
 class WiredHeadsetReceiverTest {
 
     private val context = mock<Context>()
-    private val logger = mock<LogWrapper>()
+    private val logger = mock<Logger>()
     private val wiredDeviceConnectionListener = mock<WiredDeviceConnectionListener>()
     private val wiredHeadsetReceiver = WiredHeadsetReceiver(
             context,